requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==2.4.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jiter==0.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
lizard==1.17.31
matplotlib-inline==0.1.7
more-itertools==10.7.0
nest-asyncio==1.6.0
openai==1.95.1
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.1.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
ruff==0.11.11
setuptools==80.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
tap.py==3.2.1
tenacity==8.5.0
tiktoken==0.8.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
